dev: partially restored mining dashboard fragment + fixed bug with parsed double

+++ app/src/main/java/com/donuts/bismuth/bismuthtoolbox/Data/ParsedHomeScreenData.java
@@ -42,7 +42,7 @@ public class ParsedHomeScreenData {
     private int minersInactive;
 
     @ColumnInfo(name = "miners_hashrate")
-    private double minersHashrate;
+    private int minersHashrate;
 
     @ColumnInfo(name = "block_height")
     private int blockHeight;
@@ -128,11 +128,11 @@ public class ParsedHomeScreenData {
         this.minersInactive = minersInactive;
     }
 
-    public double getMinersHashrate() {
+    public int getMinersHashrate() {
         return minersHashrate;
     }
 
-    public void setMinersHashrate(double minersHashrate) {
+    public void setMinersHashrate(int minersHashrate) {
         this.minersHashrate = minersHashrate;
     }
